create currencyConversion module, import/export module function, update function references
